correct work district and calculator

+++ app/src/main/java/com/example/kilt/data/FilterValue.kt
@@ -5,5 +5,5 @@ sealed class FilterValue {
     data class SingleValue(val value: Int) : FilterValue()
     data class ListValue(val values: List<Int>) : FilterValue()
     data class ListValue1(val values: List<String>) : FilterValue()
-    data class RangeValue(val from: Int, val to: Int) : FilterValue()
+    data class RangeValue(val from: Long, val to: Long) : FilterValue()
 }
